[INTERPRETER] * BUG: Some structure padding and alignment fixes.
[GB.DB]
* BUG: Some structure padding and alignment fixes.

[GB.DEBUG]
* BUG: Some structure padding and alignment fixes.

[GB.DRAW]
* BUG: Some structure padding and alignment fixes.

[GB.EVAL]
* BUG: Some structure padding and alignment fixes.


git-svn-id: svn://localhost/gambas/branches/2.0@2372 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/gbx/gbx_class.h
@@ -228,7 +228,7 @@ enum
 typedef
   struct _CLASS {
     struct _CLASS *class;     /*   4  Points at the 'Class' class ! */
-    int ref;                  /*   8  Reference count */
+    intptr_t ref;             /*   8  Reference count */
     int count;                /*  12  Number of instanciated objects */
     struct _CLASS *parent;    /*  16  Inherited class */
     COMPONENT *component;     /*  20  The component the class belongs to */
